Bugfixes!  Stuff works better now.
The @forward() and @continue_() decorators are now gentler and more
compatible with respect to __init__.  The rule now is:
   * in @forward() we set our own __init__ so it throws a TypeError
   * in @continue_(),
      if they set their own __init__
        we leave it alone
      else
        del __init__
This works better.  Before it was effectively doing this:
    cls.__init__ = getattr(cls, '__init__')
and that seems to upset Python's delicate stomach, somehow.

toggle_forward.py now cleans up after itself a little more.
It now deletes "forward" and "continue_" at the end of files
it adds them to.

Also, it uses the class name "_____" for the throwaway class
object, in case a file is using "_" for a significant and permanent
object name.  (Lib/argparse.py seems to be doing that.)